finish cart with redis

+++ Backend/2Sport_BE/Controllers/CartWithRedisController.cs
@@ -8,6 +8,7 @@ using _2Sport_BE.ViewModels;
 using AutoMapper;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using NuGet.Configuration;
 using System.Security.Claims;
 
 namespace _2Sport_BE.Controllers
@@ -22,12 +23,13 @@ namespace _2Sport_BE.Controllers
         private readonly IProductService _productService;
         private readonly IRedisCacheService _redisCacheService;
         private readonly IMapper _mapper;
-
+        private readonly string _cartItemsKey;
         public CartWithRedisController(IUnitOfWork unitOfWork,
                                        ICartItemService cartItemService,
                                        IWarehouseService warehouse,
                                        IProductService productService,
                                        IRedisCacheService redisCacheService,
+                                       IConfiguration configuration,
                                        IMapper mapper)
         {
             _unitOfWork = unitOfWork;
@@ -35,6 +37,7 @@ namespace _2Sport_BE.Controllers
             _productService = productService;
             _redisCacheService = redisCacheService;
             _warehouseService = warehouse;
+            _cartItemsKey = configuration.GetValue<string>("RedisKeys:CartItems");
             _mapper = mapper;
         }
 
@@ -50,7 +53,7 @@ namespace _2Sport_BE.Controllers
                 {
                     return Unauthorized();
                 }
-                var query = _redisCacheService.GetData<List<CartItem>>("CartItems");
+                var query = _redisCacheService.GetData<List<CartItem>>(_cartItemsKey);
                 //var query = await _cartItemService.GetCartItems(userId, defaultSearch.currentPage, defaultSearch.perPage);
                 if (query != null)
                 {
@@ -89,10 +92,22 @@ namespace _2Sport_BE.Controllers
         {
             try
             {
-                var cartItem = await _cartItemService.GetCartItemById(cartItemId);
+                var listCartItems = _redisCacheService.GetData<List<CartItem>>(_cartItemsKey) 
+                                                            ?? new List<CartItem>();
+                var cartItem = listCartItems.Find(_ => _.CartItemId.Equals(cartItemId));
                 if (cartItem != null)
                 {
-                    return Ok(cartItem);
+                    var result = _mapper.Map<CartItemVM>(cartItem);
+                    var product = await _unitOfWork.ProductRepository.FindAsync(cartItem.ProductId);
+                    result.ProductName = product.ProductName;
+                    result.ProductCode = product.ProductCode;
+                    result.Color = product.Color;
+                    result.Size = product.Size;
+                    result.Price = (decimal)product.Price;
+                    result.Condition = (int)product.Condition;
+                    result.ImgAvatarPath = product.ImgAvatarPath;
+                    result.RentPrice = (decimal)product.RentPrice;
+                    return Ok(result);
                 }
                 return BadRequest($"Cannot find cart item with id: {cartItemId}");
             }
@@ -125,7 +140,7 @@ namespace _2Sport_BE.Controllers
                 }
                 var quantityOfProduct = productInWarehouse.AvailableQuantity;
                 var addedCartItemQuantity = cartItemCM.Quantity;
-                var listCartItemsInCache = _redisCacheService.GetData<List<CartItem>>("CartItems") 
+                var listCartItemsInCache = _redisCacheService.GetData<List<CartItem>>(_cartItemsKey) 
                                                                         ?? new List<CartItem>();
                 var existedCartItem = listCartItemsInCache.FirstOrDefault(_ => _.UserId == userId
                                                                 && _.ProductId == (int)cartItemCM.ProductId);
@@ -140,7 +155,7 @@ namespace _2Sport_BE.Controllers
                     }
                     existedCartItem.Quantity = addedCartItemQuantity;
                     existedCartItem.Price = addedProduct.Price * addedCartItemQuantity;
-                    _redisCacheService.SetData("CartItems", listCartItemsInCache);
+                    _redisCacheService.SetData(_cartItemsKey, listCartItemsInCache);
                     return Ok(existedCartItem);
                     //addedCartItem = await _cartItemService.AddExistedCartItem(existedCartItem);
 
@@ -154,7 +169,7 @@ namespace _2Sport_BE.Controllers
                     newCartItem.CartItemId = Guid.NewGuid();
                     newCartItem.UserId = userId;
                     listCartItemsInCache.Add(newCartItem);
-                    _redisCacheService.SetData("CartItems", listCartItemsInCache);
+                    _redisCacheService.SetData(_cartItemsKey, listCartItemsInCache);
                     return Ok(newCartItem);
                     //addedCartItem = await _cartItemService.AddNewCartItem(userId, newCartItem);
                 }
@@ -167,7 +182,7 @@ namespace _2Sport_BE.Controllers
 
         [HttpPut]
         [Route("reduce-cart/{cartItemId}")]
-        public async Task<IActionResult> ReduceCart(int cartItemId)
+        public async Task<IActionResult> ReduceCart(Guid cartItemId)
         {
             try
             {
@@ -177,8 +192,22 @@ namespace _2Sport_BE.Controllers
                 {
                     return Unauthorized();
                 }
-
-                await _cartItemService.ReduceCartItem(cartItemId);
+                var listCartItems = _redisCacheService.GetData<List<CartItem>>(_cartItemsKey)
+                                                                    ?? new List<CartItem>();
+                var reducedCartItem = listCartItems.Find(_ => _.CartItemId.Equals(cartItemId));
+                if (reducedCartItem != null)
+                {
+                    var product = (await _unitOfWork.ProductRepository
+                                                        .GetAsync(_ => _.Id == reducedCartItem.ProductId))
+                                                        .FirstOrDefault();
+                    reducedCartItem.Quantity -= 1;
+                    reducedCartItem.Price -= product.Price;
+                    if (reducedCartItem.Quantity == 0)
+                    {
+                        listCartItems.Remove(reducedCartItem);
+                    }
+                    _redisCacheService.SetData(_cartItemsKey, listCartItems);
+                }
                 return Ok($"Reduce cart item with id: {cartItemId}");
             }
             catch (Exception ex)
@@ -199,14 +228,32 @@ namespace _2Sport_BE.Controllers
                 {
                     return Unauthorized();
                 }
-                var cartItem = await _cartItemService.GetCartItemById(cartItemId);
-                var quantityOfProduct = (await _warehouseService.GetWarehouseByProductId(cartItem.ProductId))
+                var listCartItems = _redisCacheService.GetData<List<CartItem>>(_cartItemsKey)
+                                                              ?? new List<CartItem>();
+                var updatedCartItem = listCartItems.Find(_ => _.CartItemId.Equals(cartItemId));
+                if (updatedCartItem == null)
+                {
+                    return NotFound("There is not cart item!");
+                }
+
+                var quantityOfProduct = (await _warehouseService.GetWarehouseByProductId(updatedCartItem.ProductId))
                         .FirstOrDefault().AvailableQuantity;
                 if (quantity > quantityOfProduct)
                 {
                     return BadRequest($"Xin lỗi! Chúng tôi chỉ còn {quantityOfProduct} sản phẩm");
                 }
-                await _cartItemService.UpdateQuantityOfCartItem(cartItemId, quantity);
+
+                var product = await _unitOfWork.ProductRepository.FindAsync(updatedCartItem.ProductId);
+                if (updatedCartItem.Quantity == 0)
+                {
+                    listCartItems.Remove(updatedCartItem);
+                }
+                else
+                {
+                    updatedCartItem.Quantity = quantity;
+                    updatedCartItem.Price = quantity * product.Price;
+                }
+                _redisCacheService.SetData(_cartItemsKey, listCartItems);
                 return Ok($"Update quantity cart item with id: {cartItemId}");
             }
             catch (Exception ex)
@@ -227,7 +274,15 @@ namespace _2Sport_BE.Controllers
                 {
                     return Unauthorized();
                 }
-
+                var listCartItems = _redisCacheService.GetData<List<CartItem>>(_cartItemsKey)
+                                                       ?? new List<CartItem>();
+                var deletedCartItem = listCartItems.Find(_ => _.CartItemId.Equals(cartItemId));
+                if (deletedCartItem == null)
+                {
+                    return NotFound("There is not cart item!");
+                } 
+                listCartItems.Remove(deletedCartItem);
+                _redisCacheService.SetData(_cartItemsKey, listCartItems);
                 await _cartItemService.DeleteCartItem(cartItemId);
                 return Ok($"Delete cart item with id: {cartItemId}");
             }
